Fork presentation into notes and performance.
My talk was getting too wordy. There needs to be a "performance"
(short, sweet, to the point) version and a "consultant" or "notes of
slide deck" version for public consumption by those unable to attend
my talk. However, it seemed inappropriate to branch between the two.
In an ideal world, I'd just mark the Markdown with some CSS-like
information to indicate which parts are the notes and which parts
are the performance so I can keep it all in one document. Given time
constraints, two separate documents seems better, and they can each
use common infrastructure (stylesheets).